+++ bigraph/src/sbrs.ml
@@ -6,7 +6,9 @@ type react = {
   (* Reactum --- rhs *)
   eta : Fun.t option;
   (* Instantiation map *)
-  rate : float; (* Reaction rate *)
+  rate : float;
+  (* Reaction rate *)
+  conds : AppCond.t list; (* Application conditions *)
 }
 
 module RT = struct
@@ -19,6 +21,8 @@ module RT = struct
   let lhs r = r.rdx
 
   let rhs r = r.rct
+
+  let conds r = r.conds
 
   let l r = r.rate
 
@@ -37,11 +41,11 @@ module RT = struct
 
   let string_of_label = Printf.sprintf "%-3g"
 
-  let parse ~name ~lhs ~rhs r eta =
-    { name; rdx = lhs; rct = rhs; eta; rate = r }
+  let parse ~name ~lhs ~rhs ?conds:(c = []) r eta =
+    { name; rdx = lhs; rct = rhs; eta; rate = r; conds = c }
 
   let step b rules =
-    RrType.gen_step b rules merge_occ ~lhs ~rhs ~label:l ~map
+    RrType.gen_step b rules merge_occ ~lhs ~rhs ~label:l ~map ~conds
 
   let random_step b rules =
     (* Sort transitions by rate *)